requirements.txt
asgiref==3.3.1
Django==3.1.4
django-admin-list-filter-dropdown==1.0.3
django-modeladmin-reorder==0.3.1
django-searchable-select==1.5.0
django-smart-selects==1.5.9
djangorestframework==3.12.2
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
